requirements.txt
# Existing dependencies
openai
python-dotenv
fastapi
uvicorn
redis
mysql-connector-python

# Authentication dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0 